fix: three-dots anim is squished to 0px height
AnimatedSprite2D is not a child of the Control node, so the parent
MarginContainer cannot calculate its size correctly. When the scroll-area is
filled (a scrollbar is needed), its height is squeezed down up to a minimum,
defined by the size of its children.

However, as the DottedAnimBox has no Control children, its minimum size
is 0x0. Thus, when the phone screen is full, tehe DottedAnimBox is
squeezed to size 0 on the Y-axis, effectively making it invisible.

Fixed by setting the minimum size of the DottedAnimBox to the desired
size of the animated sprite.