fix: no dot anim immediately after a response
The time_before_typing was not being started in select_response. It was
only kicked off after the bot sends a messsage (in message_timer timeout),
and immediately in _ready().

Therefore, the first message after the player selects a response would
not play the dot animation.

Fixed by adding in the missing .start() to select_response.